refactor(media): split the youtube host into focused modules
media.ts had grown to ~950 lines holding four unrelated concerns: the
vendor API surface, source parsing, URL building, and the host itself.
Split along those seams, following how `hls-js` and `google-cast` are laid
out.

- iframe-api.ts: vendor typings, the script loader, and the state and error
  code tables that only mean something to YouTube
- source.ts: engine config, source types, parsing, and embed URL building
- props.ts: props and defaults, shared by the host and the URL builder
  without either importing the other
- media.ts: the host class alone

Three helpers this PR had copied from the Vimeo host move to dom/utils and
are now shared: serializeEmbedParams, createTimeRange, and
createPublicPromise. Vimeo uses them too, so the duplicates are gone rather
than doubled.

The barrel re-exports the same names as before. iframe-api is internal
apart from the player typings the `engine` getter surfaces.
