+++ packages/media/src/dom/utils/index.ts
@@ -1,3 +1,4 @@
+export { serializeEmbedParams } from './embed-params';
 export {
   addMediaComponent,
   getMediaComponents,
@@ -5,3 +6,5 @@ export {
   getMediaProp,
   setMediaProp,
 } from './media-components';
+export { createPublicPromise, type PublicPromise } from './public-promise';
+export { createTimeRange } from './time-ranges';
