+++ packages/media/src/dom/youtube/index.ts
@@ -1,1 +1,6 @@
+// The iframe API module is internal apart from the player typings, which the
+// `engine` getter surfaces.
+export type { YouTubeApi, YouTubePlayerApi } from './iframe-api';
 export * from './media';
+export * from './props';
+export * from './source';
